requirements.txt
Flask==2.2.2
Flask_Cors==3.0.10
Flask_JWT_Extended==4.4.4
flask_sqlalchemy==3.0.2
jusText==3.0.0
nltk==3.8
numpy==1.21.5
openai==0.27.2
py-bcrypt==0.4
PyYAML==6.0
requests
scikit_learn==1.2.0
textract==1.6.5
trafilatura==1.4.1
pyopenssl
langdetect==1.0.9
waitress==2.1.2
cachelib==0.10.2
tweepy==4.13.0
gunicorn==20.1.0
sendgrid==6.9.7
Flask-Migrate==4.0.4
notion-client==2.0.0
beautifulsoup4==4.8.2
retrying==1.3.4
timeout_decorator==0.5.0
youtube_transcript_api==0.5.0
retry==0.9.2